pyproject.toml
[tool.poetry]
name = "aiopulsar"
version = "0.0.0"
description = "Asynchronous wrapper for the python pulsar-client."
authors = ["Mads Hansen Baattrup <mads@baattrup.com>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"
pulsar-client = "^2.10.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.950"
flake8 = "^4.0.1"
poetry-githooks = "^2.0.0"
pytest-asyncio = "^0.18.3"

[tool.black]
skip-string-normalization = true

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = [
    "tests",
]

[tool.githooks]
pre-commit = """black . &&
                echo "black success" &&
                flake8 . &&
                echo "flake8 success" &&
                mypy . &&
                echo "mypy success"
            """

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
